some updates

+++ fire_risk_classifier/dataclasses/house_point.py
@@ -27,7 +27,7 @@ class HousePoint:
         self.longitude = float(longitude)
         self.fire_risk = FireRisk.from_style_url(style_url, name)
 
-    def get_bbox(self, radius: int) -> BoundingBox:
+    def get_bbox(self, radius: float) -> BoundingBox:
         # Earth's radius in meters
         earth_radius = 6371000
